requirements.txt
azure-functions
requests>=2.31.0
urllib3>=2.0.0
pyodbc
python-dateutil